+++ lib/kitchen/driver/ec2.rb
@@ -25,6 +25,7 @@ require_relative "aws/client"
 require_relative "aws/instance_generator"
 require_relative "aws/standard_platform"
 require_relative "aws/standard_platform/amazon"
+require_relative "aws/standard_platform/amazon2"
 require_relative "aws/standard_platform/centos"
 require_relative "aws/standard_platform/debian"
 require_relative "aws/standard_platform/rhel"
@@ -32,13 +33,12 @@ require_relative "aws/standard_platform/fedora"
 require_relative "aws/standard_platform/freebsd"
 require_relative "aws/standard_platform/ubuntu"
 require_relative "aws/standard_platform/windows"
+require "aws-sdk-ec2"
 require "aws-sdk-core/waiters/errors"
 require "retryable"
 require "time"
 require "etc"
 require "socket"
-
-Aws.eager_autoload!
 
 module Kitchen
 
@@ -62,6 +62,7 @@ module Kitchen
       default_config :ebs_optimized,      false
       default_config :security_group_ids, nil
       default_config :security_group_filter, nil
+      default_config :security_group_cidr_ip, "0.0.0.0/0"
       default_config :tags, "created-by" => "test-kitchen"
       default_config :user_data do |driver|
         if driver.windows_os?
@@ -94,9 +95,6 @@ module Kitchen
 
       def initialize(*args, &block)
         super
-        # AWS Ruby SDK loading isn't thread safe, so as soon as we know we're
-        # going to use EC2, autoload it. Seems to have been fixed in Ruby 2.3+
-        ::Aws.eager_autoload! unless RUBY_VERSION.to_f >= 2.3
       end
 
       def self.validation_warn(driver, old_key, new_key)
@@ -250,15 +248,15 @@ module Kitchen
         # Tagging an instance is possible before volumes are attached. Tagging the volumes after
         # instance creation is consistent.
         Retryable.retryable(
-          :tries => 10,
-          :sleep => lambda { |n| [2**n, 30].min },
-          :on => ::Aws::EC2::Errors::InvalidInstanceIDNotFound
+          tries: 10,
+          sleep: lambda { |n| [2**n, 30].min },
+          on: ::Aws::EC2::Errors::InvalidInstanceIDNotFound
         ) do |r, _|
           info("Attempting to tag the instance, #{r} retries")
           tag_server(server)
 
           # Get information about the AMI (image) used to create the image.
-          image_data = ec2.client.describe_images({ :image_ids => [server.image_id] })[0][0]
+          image_data = ec2.client.describe_images({ image_ids: [server.image_id] })[0][0]
 
           state[:server_id] = server.id
           info("EC2 instance <#{state[:server_id]}> created.")
@@ -301,7 +299,7 @@ module Kitchen
           if state[:spot_request_id]
             debug("Deleting spot request <#{state[:server_id]}>")
             ec2.client.cancel_spot_instance_requests(
-              :spot_instance_request_ids => [state[:spot_request_id]]
+              spot_instance_request_ids: [state[:spot_request_id]]
             )
             state.delete(:spot_request_id)
           end
@@ -431,7 +429,7 @@ module Kitchen
         state[:spot_request_id] = spot_request_id
         ec2.client.wait_until(
           :spot_instance_request_fulfilled,
-          :spot_instance_request_ids => [spot_request_id]
+          spot_instance_request_ids: [spot_request_id]
         ) do |w|
           w.max_attempts = config[:retryable_tries]
           w.delay = config[:retryable_sleep]
@@ -447,9 +445,9 @@ module Kitchen
       def create_spot_request
         request_duration = config[:retryable_tries] * config[:retryable_sleep]
         request_data = {
-          :spot_price => config[:spot_price].to_s,
-          :launch_specification => instance_generator.ec2_instance_data,
-          :valid_until => Time.now + request_duration,
+          spot_price: config[:spot_price].to_s,
+          launch_specification: instance_generator.ec2_instance_data,
+          valid_until: Time.now + request_duration,
         }
         if config[:block_duration_minutes]
           request_data[:block_duration_minutes] = config[:block_duration_minutes]
@@ -465,19 +463,19 @@ module Kitchen
             # we convert the value to a string because
             # nils should be passed as an empty String
             # and Integers need to be represented as Strings
-            { :key => k, :value => v.to_s }
+            { key: k, value: v.to_s }
           end
-          server.create_tags(:tags => tags)
+          server.create_tags(tags: tags)
         end
       end
 
       def tag_volumes(server)
         if config[:tags] && !config[:tags].empty?
           tags = config[:tags].map do |k, v|
-            { :key => k, :value => v.to_s }
+            { key: k, value: v.to_s }
           end
           server.volumes.each do |volume|
-            volume.create_tags(:tags => tags)
+            volume.create_tags(tags: tags)
           end
         end
       end
@@ -491,8 +489,8 @@ module Kitchen
           described_volume_count = 0
           ready_volume_count = 0
           if aws_instance.exists?
-            described_volume_count = ec2.client.describe_volumes(:filters => [
-              { :name => "attachment.instance-id", :values => ["#{state[:server_id]}"] }]
+            described_volume_count = ec2.client.describe_volumes(filters: [
+              { name: "attachment.instance-id", values: ["#{state[:server_id]}"] }]
               ).volumes.length
             aws_instance.volumes.each { ready_volume_count += 1 }
           end
@@ -512,6 +510,11 @@ module Kitchen
           ready = aws_instance.exists? &&
             aws_instance.state.name == "running" &&
             hostname != "0.0.0.0"
+
+          if ready && ( hostname.nil? || hostname == "" )
+            debug("Unable to detect hostname using interface_type #{config[:interface]}. Fallback to ordered mapping")
+            state[:hostname] = hostname(aws_instance, nil)
+          end
           if ready && windows_os?
             output = server.console_output.output
             unless output.nil?
@@ -535,9 +538,9 @@ module Kitchen
         begin
           with_request_limit_backoff(state) do
             server.wait_until(
-              :max_attempts => config[:retryable_tries],
-              :delay => config[:retryable_sleep],
-              :before_attempt => wait_log,
+              max_attempts: config[:retryable_tries],
+              delay: config[:retryable_sleep],
+              before_attempt: wait_log,
               &block
             )
           end
@@ -552,7 +555,7 @@ module Kitchen
       def fetch_windows_admin_password(server, state)
         wait_with_destroy(server, state, "to fetch windows admin password") do |aws_instance|
           enc = server.client.get_password_data(
-            :instance_id => state[:server_id]
+            instance_id: state[:server_id]
           ).password_data
           # Password data is blank until password is available
           !enc.nil? && !enc.empty?
@@ -588,7 +591,7 @@ module Kitchen
           "public" => "public_ip_address",
           "private" => "private_ip_address",
           "private_dns" => "private_dns_name",
-        }
+        }.freeze
 
       #
       # Lookup hostname of provided server. If interface_type is provided use
@@ -632,19 +635,18 @@ module Kitchen
       def default_windows_user_data
         base_script = Kitchen::Util.outdent!(<<-EOH)
 	$OSVersion = (get-itemproperty -Path "HKLM:\\SOFTWARE\\Microsoft\\Windows NT\\CurrentVersion" -Name ProductName).ProductName
-	If($OSVersion.contains('2016'))
-	{
-      	$logfile='C:\\ProgramData\\Amazon\\EC2-Windows\\Launch\\Log\\kitchen-ec2.log'
-	# EC2Launch doesn't init extra disks by default
-	C:\\ProgramData\\Amazon\\EC2-Windows\\Launch\\Scripts\\InitializeDisks.ps1
-	}
-	Else
-	{
-	$logfile='C:\\Program Files\\Amazon\\Ec2ConfigService\\Logs\\kitchen-ec2.log'
+  If($OSVersion.contains('2016') -Or $OSVersion -eq 'Windows Server Datacenter') {
+    New-Item -ItemType Directory -Force -Path 'C:\\ProgramData\\Amazon\\EC2-Windows\\Launch\\Log'
+    $logfile='C:\\ProgramData\\Amazon\\EC2-Windows\\Launch\\Log\\kitchen-ec2.log'
+    # EC2Launch doesn't init extra disks by default
+    C:\\ProgramData\\Amazon\\EC2-Windows\\Launch\\Scripts\\InitializeDisks.ps1
+  } Else {
+     New-Item -ItemType Directory -Force -Path 'C:\\Program Files\\Amazon\\Ec2ConfigService\\Logs'
+     $logfile='C:\\Program Files\\Amazon\\Ec2ConfigService\\Logs\\kitchen-ec2.log'
   }
 
         # Logfile fail-safe in case the directory does not exist
-        New-Item $logfile -Force
+        New-Item $logfile -Type file -Force
 
         # Allow script execution
         Set-ExecutionPolicy -ExecutionPolicy RemoteSigned -Force
@@ -662,7 +664,7 @@ module Kitchen
         #Firewall Config
         & netsh advfirewall firewall set rule name="Windows Remote Management (HTTP-In)" profile=public protocol=tcp localport=5985 remoteip=localsubnet new remoteip=any  >> $logfile
         Set-ItemProperty -Name LocalAccountTokenFilterPolicy -Path HKLM:\\software\\Microsoft\\Windows\\CurrentVersion\\Policies\\system -Value 1
-	EOH
+        EOH
 
         # Preparing custom static admin user if we defined something other than Administrator
         custom_admin_script = ""
@@ -706,8 +708,8 @@ module Kitchen
       end
 
       def image_info(image)
-        root_device = image.block_device_mappings.
-          find { |b| b.device_name == image.root_device_name }
+        root_device = image.block_device_mappings
+          .find { |b| b.device_name == image.root_device_name }
         volume_type = " #{root_device.ebs.volume_type}" if root_device && root_device.ebs
 
         " Architecture: #{image.architecture}," \
@@ -758,12 +760,12 @@ module Kitchen
         ec2.client.authorize_security_group_ingress(
           group_id: state[:auto_security_group_id],
           # Allow SSH and WinRM (both plain and TLS).
-          ip_permissions: [22, 5985, 5986].map do |port|
+          ip_permissions: [22, 3389, 5985, 5986].map do |port|
             {
               ip_protocol: "tcp",
               from_port: port,
               to_port: port,
-              ip_ranges: [{ cidr_ip: "0.0.0.0/0" }],
+              ip_ranges: [{ cidr_ip: config[:security_group_cidr_ip] }],
             }
           end
         )