+++ CHANGELOG.md
@@ -1,5 +1,45 @@
 # Change Log
 
+## [v2.4.0](https://github.com/test-kitchen/kitchen-ec2/tree/v2.4.0) (2018-12-20)
+[Full Changelog](https://github.com/test-kitchen/kitchen-ec2/compare/v2.3.4..v2.4.0)
+
+- Don't ship spec files in the gem artifact
+- Support Amazon Linux 2.0 image searching. Use the platform 'amazon2'
+- Support Windows Server 1709 and 1803 image searching
+
+## [v2.3.4](https://github.com/test-kitchen/kitchen-ec2/tree/v2.3.4) (2018-12-04)
+[Full Changelog](https://github.com/test-kitchen/kitchen-ec2/compare/v2.3.3...v2.3.4)
+
+- Don't ship the changelog in the gem
+
+## [v2.3.3](https://github.com/test-kitchen/kitchen-ec2/tree/v2.3.3) (2018-11-28)
+[Full Changelog](https://github.com/test-kitchen/kitchen-ec2/compare/v2.3.2...v2.3.3)
+
+**Merged pull requests:**
+
+- Adding support for arm64 architecture [\#433]
+
+## [v2.3.2](https://github.com/test-kitchen/kitchen-ec2/tree/v2.3.2) (2018-11-28)
+[Full Changelog](https://github.com/test-kitchen/kitchen-ec2/compare/v2.3.1...v2.3.2)
+
+**Fixed Bugs:**
+
+- fix x86_64 architecture default for image search (fixes new arm64 arch appearing instead) [\#432]
+
+## [v2.3.1](https://github.com/test-kitchen/kitchen-ec2/tree/v2.3.1) (2018-10-19)
+[Full Changelog](https://github.com/test-kitchen/kitchen-ec2/compare/v2.3.0...v2.3.1)
+
+**Fixed Bugs:**
+
+- windows2012-r2 hanging on userdata.ps1 in kitchen-ec2 2.3.0 [\#424]
+
+## [v2.3.0](https://github.com/test-kitchen/kitchen-ec2/tree/v2.3.0) (2018-10-05)
+[Full Changelog](https://github.com/test-kitchen/kitchen-ec2/compare/v2.2.2...v2.3.0)
+
+- Add port 3389 (RDP) to the automatically generated security group
+- Fix logfile creation on Windows to not fail if the directory doesn't exist
+- The gem no longer ships with test deps so we can slim the install size
+
 ## [v2.2.2](https://github.com/test-kitchen/kitchen-ec2/tree/v2.2.2) (2018-06-11)
 [Full Changelog](https://github.com/test-kitchen/kitchen-ec2/compare/v2.2.1...v2.2.2)